rename pkg ltresolver to amifamily
