+++ pkg/cloudprovider/aws/amifamily/bootstrap/bootstrap.go
