+++ pkg/cloudprovider/aws/amifamily/bootstrap/bottlerocket.go
