+++ pkg/cloudprovider/aws/amifamily/bootstrap/eksbootstrap.go
