+++ pkg/cloudprovider/aws/amifamily/bottlerocket.go
@@ -12,7 +12,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package ltresolver
+package amifamily
 
 import (
 	"fmt"
@@ -23,8 +23,8 @@ import (
 
 	"github.com/aws/karpenter/pkg/apis/provisioning/v1alpha5"
 	"github.com/aws/karpenter/pkg/cloudprovider"
+	"github.com/aws/karpenter/pkg/cloudprovider/aws/amifamily/bootstrap"
 	"github.com/aws/karpenter/pkg/cloudprovider/aws/apis/v1alpha1"
-	"github.com/aws/karpenter/pkg/cloudprovider/aws/ltresolver/bootstrap"
 )
 
 type Bottlerocket struct {
