+++ pkg/cloudprovider/aws/amifamily/resolver.go
@@ -12,7 +12,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package ltresolver
+package amifamily
 
 import (
 	"context"
@@ -26,8 +26,8 @@ import (
 
 	"github.com/aws/karpenter/pkg/apis/provisioning/v1alpha5"
 	"github.com/aws/karpenter/pkg/cloudprovider"
+	"github.com/aws/karpenter/pkg/cloudprovider/aws/amifamily/bootstrap"
 	"github.com/aws/karpenter/pkg/cloudprovider/aws/apis/v1alpha1"
-	"github.com/aws/karpenter/pkg/cloudprovider/aws/ltresolver/bootstrap"
 )
 
 var defaultEBS = v1alpha1.BlockDevice{
@@ -55,7 +55,7 @@ type Options struct {
 	Labels            map[string]string `hash:"ignore"`
 }
 
-// ResolvedTemplate holds the dynamically generated launch tempalte parameters
+// ResolvedTemplate holds the dynamically generated launch template parameters
 type ResolvedTemplate struct {
 	*Options
 	UserData            bootstrap.Bootstrapper
